PHP Commenced
The overall design is still a big factor here, I'm not even close to finished. However, I started some of the PHP atleast getting the structure of things in order, including files etc. The database is the next step once I've finished the overall design of the pages and navigation.